Add fuzz target for using abstract policies to test TPE authorization
